+++ lib/app/app_info.dart
@@ -1,6 +1,6 @@
 class AppInfo {
   static const String appName = 'Music Music';
-  static const String appVersion = '1.2.0+1';
+  static const String appVersion = '1.3.0+1';
   static const String packageDescription =
       'Aplicativo de musica local com player premium, biblioteca inteligente e playlists.';
 }